Menu entry swapper changes
