+++ meteor-client/src/main/java/meteor/plugins/menuentryswapper/MenuEntrySwapperPlugin.java
@@ -33,9 +33,7 @@ import static meteor.plugins.menuentryswapper.MenuEntrySwapperConfig.DesertAmule
 import static meteor.plugins.menuentryswapper.MenuEntrySwapperConfig.KaramjaGlovesMode;
 import static meteor.plugins.menuentryswapper.MenuEntrySwapperConfig.MorytaniaLegsMode;
 import static meteor.plugins.menuentryswapper.MenuEntrySwapperConfig.RadasBlessingMode;
-
 import com.google.common.annotations.VisibleForTesting;
-import com.google.common.base.Splitter;
 import com.google.common.collect.ArrayListMultimap;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.LinkedHashMultimap;
@@ -43,13 +41,10 @@ import com.google.common.collect.Multimap;
 import com.google.common.collect.Multimaps;
 import com.google.common.collect.Sets;
 import com.google.inject.Provides;
-import java.awt.event.KeyEvent;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
-import java.util.Objects;
 import java.util.Set;
 import java.util.function.Predicate;
 import java.util.function.Supplier;
@@ -71,13 +66,13 @@ import meteor.plugins.PluginDescriptor;
 import meteor.plugins.menuentryswapper.util.BurningAmuletMode;
 import meteor.plugins.menuentryswapper.util.BuyMode;
 import meteor.plugins.menuentryswapper.util.CombatBraceletMode;
+import meteor.plugins.menuentryswapper.util.ConstructionCapeMode;
+import meteor.plugins.menuentryswapper.util.CraftingCapeMode;
+import meteor.plugins.menuentryswapper.util.DigsitePendantMode;
 import meteor.plugins.menuentryswapper.util.DuelingRingMode;
 import meteor.plugins.menuentryswapper.util.FairyRingMode;
 import meteor.plugins.menuentryswapper.util.GEItemCollectMode;
 import meteor.plugins.menuentryswapper.util.GamesNecklaceMode;
-import meteor.plugins.menuentryswapper.util.ConstructionCapeMode;
-import meteor.plugins.menuentryswapper.util.CraftingCapeMode;
-import meteor.plugins.menuentryswapper.util.DigsitePendantMode;
 import meteor.plugins.menuentryswapper.util.GloryMode;
 import meteor.plugins.menuentryswapper.util.HouseAdvertisementMode;
 import meteor.plugins.menuentryswapper.util.HouseMode;
@@ -92,7 +87,6 @@ import meteor.plugins.menuentryswapper.util.SkillsNecklaceMode;
 import meteor.plugins.menuentryswapper.util.Swap;
 import meteor.plugins.menuentryswapper.util.XericsTalismanMode;
 import meteor.util.Text;
-import meteor.util.WildcardMatcher;
 import net.runelite.api.Client;
 import net.runelite.api.GameState;
 import net.runelite.api.ItemComposition;
@@ -102,13 +96,11 @@ import net.runelite.api.MenuEntry;
 import net.runelite.api.NPC;
 import net.runelite.api.Varbits;
 import net.runelite.api.events.ClientTick;
-import net.runelite.api.events.FocusChanged;
 import net.runelite.api.events.MenuEntryAdded;
 import net.runelite.api.events.MenuOpened;
 import net.runelite.api.events.MenuOptionClicked;
 import net.runelite.api.events.PostItemComposition;
 import net.runelite.api.events.WidgetMenuOptionClicked;
-import net.runelite.api.widgets.Widget;
 import net.runelite.api.widgets.WidgetID;
 import net.runelite.api.widgets.WidgetInfo;
 import org.apache.commons.lang3.ArrayUtils;
@@ -118,16 +110,7 @@ import org.apache.commons.lang3.ArrayUtils;
     description = "Change the default option that is displayed when hovering over objects",
     tags = {"npcs", "inventory", "items", "objects"}
 )
-public class MenuEntrySwapperPlugin extends Plugin implements KeyListener {
-
-  private static final Set<MenuAction> EXAMINE_TYPES;
-  private static final Splitter SPLITTER;
-  private final ArrayList<EntryFromConfig> customSwaps = new ArrayList();
-  private final ArrayList<MenuEntrySwapperPlugin.EntryFromConfig> shiftCustomSwaps = new ArrayList();
-  private final ArrayList<MenuEntrySwapperPlugin.EntryFromConfig> removeOptions = new ArrayList();
-  private final ArrayList<MenuEntrySwapperPlugin.EntryFromConfig> bankCustomSwaps = new ArrayList();
-  private final ArrayList<MenuEntrySwapperPlugin.EntryFromConfig> shiftBankCustomSwaps = new ArrayList();
-  private boolean holdingShift;
+public class MenuEntrySwapperPlugin extends Plugin {
 
   private MenuEntry[] menuEntries;
   List<String> targetList;
@@ -211,9 +194,12 @@ public class MenuEntrySwapperPlugin extends Plugin implements KeyListener {
   @Inject
   private MenuManager menuManager;
   @Inject
-  private KeyManager keyManager;
-  @Inject
   private ItemManager itemManager;
+  @Inject
+  private zMenuEntryPlugin customswaps;
+  @Inject
+  private KeyManager keyManager;
+
   @Getter
   private boolean configuringShiftClick = false;
 
@@ -231,6 +217,8 @@ public class MenuEntrySwapperPlugin extends Plugin implements KeyListener {
 
   @Override
   public void startup() {
+    eventBus.register(customswaps);
+    keyManager.registerKeyListener(customswaps,this.getClass());
     if (config.shiftClickCustomization()) {
       enableCustomization();
     }
@@ -242,19 +230,6 @@ public class MenuEntrySwapperPlugin extends Plugin implements KeyListener {
     }
     setCastOptions(true);
     loadSwaps();
-
-    holdingShift = false;
-    keyManager.registerKeyListener(this, getClass());
-    customSwaps.clear();
-    parseConfigToList(config.customSwapsString(), customSwaps);
-    shiftCustomSwaps.clear();
-    parseConfigToList(config.shiftCustomSwapsString(), shiftCustomSwaps);
-    removeOptions.clear();
-    parseConfigToList(config.removeOptionsString(), removeOptions);
-    bankCustomSwaps.clear();
-    parseConfigToList(config.bankCustomSwapsString(), bankCustomSwaps);
-    shiftBankCustomSwaps.clear();
-    parseConfigToList(config.bankShiftCustomSwapsString(), shiftBankCustomSwaps);
   }
 
   @Override
@@ -266,9 +241,6 @@ public class MenuEntrySwapperPlugin extends Plugin implements KeyListener {
     if (client.getGameState() == GameState.LOGGED_IN) {
       resetCastOptions();
     }
-
-    holdingShift = false;
-    keyManager.unregisterKeyListener(this);
   }
 
   @VisibleForTesting
@@ -579,29 +551,6 @@ public class MenuEntrySwapperPlugin extends Plugin implements KeyListener {
           resetCastOptions();
         }
     }
-    if (event.getGroup().equals(MenuEntrySwapperConfig.GROUP))
-      switch (event.getKey()) {
-        case "customSwapsStr":
-          this.customSwaps.clear();
-          parseConfigToList(this.config.customSwapsString(), this.customSwaps);
-          break;
-        case "shiftCustomSwapsStr":
-          this.shiftCustomSwaps.clear();
-          parseConfigToList(this.config.shiftCustomSwapsString(), this.shiftCustomSwaps);
-          break;
-        case "removeOptionsStr":
-          this.removeOptions.clear();
-          parseConfigToList(this.config.removeOptionsString(), this.removeOptions);
-          break;
-        case "bankCustomSwapsStr":
-          this.bankCustomSwaps.clear();
-          parseConfigToList(this.config.bankCustomSwapsString(), this.bankCustomSwaps);
-          break;
-        case "bankShiftCustomSwapsStr":
-          this.shiftBankCustomSwaps.clear();
-          parseConfigToList(this.config.bankShiftCustomSwapsString(), this.shiftBankCustomSwaps);
-          break;
-      }
   }
 
 
@@ -757,43 +706,6 @@ public class MenuEntrySwapperPlugin extends Plugin implements KeyListener {
 
     menuEntries = client.getMenuEntries();
     swapConstructionMenu(menuEntries);
-    if (!isBankInterfaceClosed() && menuEntryAdded.getIdentifier() <= 2) {
-      MenuEntry[] menuEntries = client.getMenuEntries();
-      int entryIndex = -1;
-      int priority = -1;
-
-      for (int i = 0; i < menuEntries.length; ++i) {
-        MenuEntry entry = menuEntries[i];
-        String target = Text.removeTags(entry.getTarget()).toLowerCase();
-        String option = Text.removeTags(entry.getOption()).toLowerCase();
-        EntryFromConfig entryFromConfig = new EntryFromConfig(option, target);
-        if (holdingShift && config.shiftCustomSwapsToggle()) {
-          int index = indexOfEntry(shiftBankCustomSwaps, entryFromConfig, menuEntries);
-          if (index > priority) {
-            entryIndex = i;
-            priority = index;
-          }
-        }
-        else if (config.customSwapsToggle()) {
-          int index = indexOfEntry(bankCustomSwaps, entryFromConfig, menuEntries);
-          if (index > priority) {
-            entryIndex = i;
-            priority = index;
-          }
-        }
-      }
-
-      if (entryIndex >= 0) {
-        MenuEntry target = menuEntries[entryIndex];
-        int opId = target.getIdentifier();
-        int actionId = opId >= 6 ? MenuAction.CC_OP_LOW_PRIORITY.getId() : MenuAction.CC_OP.getId();
-        if (menuEntryAdded.getType() == MenuAction.CC_OP.getId() && (menuEntryAdded.getIdentifier() == 1
-            || menuEntryAdded.getIdentifier() == 2))
-        {
-          specialSwap(actionId, opId);
-        }
-      }
-    }
 
     if (!config.getEasyConstruction())
     {
@@ -914,54 +826,8 @@ public class MenuEntrySwapperPlugin extends Plugin implements KeyListener {
   public void onClientTick(ClientTick clientTick) {
     // The menu is not rebuilt when it is open, so don't swap or else it will
     // repeatedly swap entries
-    if (client.getGameState() == GameState.LOGGED_IN && !client.isMenuOpen() && isBankInterfaceClosed()){
+    if (client.getGameState() == GameState.LOGGED_IN && !client.isMenuOpen()){
       MenuEntry[] menuEntries = client.getMenuEntries();
-      if (config.removeOptionsToggle()) {
-        menuEntries = filterEntries(menuEntries);
-        client.setMenuEntries(menuEntries);
-      }
-
-      int entryIndex = -1;
-      int priority = -1;
-
-      int index;
-      for (int i = 0; i < menuEntries.length; ++i) {
-        MenuEntry entry = menuEntries[i];
-        String target = Text.standardize(Text.removeTags(entry.getTarget()));
-        String option = Text.standardize(Text.removeTags(entry.getOption()));
-        EntryFromConfig entryFromConfig = new EntryFromConfig(
-            option, target);
-        if (holdingShift && config.shiftCustomSwapsToggle()) {
-          index = indexOfEntry(shiftCustomSwaps, entryFromConfig, menuEntries);
-          if (index > priority) {
-            entryIndex = i;
-            priority = index;
-          }
-        }
-        else if (config.customSwapsToggle()) {
-          index = indexOfEntry(customSwaps, entryFromConfig, menuEntries);
-          if (index > priority) {
-            entryIndex = i;
-            priority = index;
-          }
-        }
-      }
-
-      if (entryIndex >= 0) {
-        MenuEntry target = menuEntries[entryIndex];
-        int targetId = target.getIdentifier();
-        int targetType = target.getType();
-        for (MenuEntry menuEntry : menuEntries) {
-          if (menuEntry.getType() < target.getType())
-            menuEntry.setType(menuEntry.getType() + 2000);
-        }
-        if (targetId >= 6 && targetId <= 9 && targetType == MenuAction.CC_OP_LOW_PRIORITY.getId())
-          target.setType(MenuAction.CC_OP.getId());
-        MenuEntry first = menuEntries[menuEntries.length - 1];
-        menuEntries[menuEntries.length - 1] = menuEntries[entryIndex];
-        menuEntries[entryIndex] = first;
-        this.client.setMenuEntries(menuEntries);
-        }
 
         // Build option map for quick lookup in findIndex
         int idx = 0;
@@ -1297,202 +1163,5 @@ public class MenuEntrySwapperPlugin extends Plugin implements KeyListener {
       client.setHideFriendCastOptions(false);
       client.setHideClanmateCastOptions(false);
     });
-  }
-
-  static class EntryFromConfig {
-
-    private String option;
-    private String target;
-    private String topOption;
-    private String topTarget;
-
-    EntryFromConfig(String option, String target) {
-      this(option, target, null, null);
-    }
-
-    public boolean equals(Object other) {
-      if (!(other instanceof MenuEntrySwapperPlugin.EntryFromConfig)) {
-        return false;
-      } else {
-        return  option.equals(((MenuEntrySwapperPlugin.EntryFromConfig) other).option)
-            &&  target.equals(((MenuEntrySwapperPlugin.EntryFromConfig) other).target)
-            &&  topOption.equals(((MenuEntrySwapperPlugin.EntryFromConfig) other).topOption)
-            &&  topTarget.equals(((MenuEntrySwapperPlugin.EntryFromConfig) other).topTarget);
-      }
-    }
-
-    public int hashCode() {
-      return Objects.hash( option,  target,  topOption,  topTarget);
-    }
-
-    public String getOption() {
-      return option;
-    }
-
-    public String getTarget() {
-      return target;
-    }
-
-    public String getTopOption() {
-      return topOption;
-    }
-
-    public String getTopTarget() {
-      return topTarget;
-    }
-
-    public void setOption(String option) {
-      this.option = option;
-    }
-
-    public void setTarget(String target) {
-      this.target = target;
-    }
-
-    public void setTopOption(String topOption) {
-      this.topOption = topOption;
-    }
-
-    public void setTopTarget(String topTarget) {
-       this.topTarget = topTarget;
-    }
-
-    public String toString() {
-      return "MenuEntrySwapperPlugin.EntryFromConfig(option=" + getOption() + ", target=" + getTarget() + ", topOption=" + getTopOption() + ", topTarget=" + getTopTarget() + ")";
-    }
-
-    public EntryFromConfig(String option, String target, String topOption, String topTarget) {
-       this.option = option;
-       this.target = target;
-       this.topOption = topOption;
-       this.topTarget = topTarget;
-    }
-  }
-
-  private void parseConfigToList(String value, List<MenuEntrySwapperPlugin.EntryFromConfig> set) {
-    List<String> strList = SPLITTER.splitToList(value);
-
-    for (String str : strList) {
-      String[] stringList = str.split(",");
-      if (stringList.length > 1) {
-        String option = stringList[0].toLowerCase().trim();
-        String target = stringList[1].toLowerCase().trim();
-        String topOption = null;
-        String topTarget = null;
-        if (stringList.length == 4) {
-          topOption = stringList[2].toLowerCase().trim();
-          topTarget = stringList[3].toLowerCase().trim();
-        }
-
-        EntryFromConfig entryFromConfig = new EntryFromConfig(
-            option, target, topOption, topTarget);
-        set.add(entryFromConfig);
-      }
-    }
-  }
-
-  private static int topEntryIndex(MenuEntry[] entries) {
-    for (int i = entries.length - 1; i >= 0; --i) {
-      int type = entries[i].getType();
-      if (!EXAMINE_TYPES.contains(MenuAction.of(type))) {
-        return i;
-      }
-    }
-
-    return entries.length - 1;
-  }
-
-  private static int indexOfEntry(List<MenuEntrySwapperPlugin.EntryFromConfig> configEntries,
-      MenuEntrySwapperPlugin.EntryFromConfig entryFromConfig, MenuEntry[] entries) {
-    int topEntryIndex = topEntryIndex(entries);
-    MenuEntry topEntry = entries[topEntryIndex];
-    String target = Text.removeTags(topEntry.getTarget()).toLowerCase();
-    String option = Text.removeTags(topEntry.getOption()).toLowerCase();
-
-    for (int i = 0; i < configEntries.size(); ++i) {
-      EntryFromConfig _configEntry = configEntries.get(i);
-      if ((_configEntry.option.equals(entryFromConfig.option) || WildcardMatcher
-          .matches(_configEntry.option, entryFromConfig.option)) && (
-          _configEntry.target.equals(entryFromConfig.target) || WildcardMatcher
-              .matches(_configEntry.target, entryFromConfig.target))) {
-        boolean a = _configEntry.topOption == null;
-        boolean b = _configEntry.topTarget == null;
-        Supplier<Boolean> c = () -> (_configEntry.topOption.equals(option) || WildcardMatcher.matches(_configEntry.topOption, option));
-        Supplier<Boolean> d = () -> (_configEntry.topTarget.equals(target) || WildcardMatcher.matches(_configEntry.topTarget, target));
-        if (a || b || (c.get() && d.get()))
-          return i;
-      }
-    }
-    return -1;
-  }
-
-  private MenuEntry[] filterEntries(MenuEntry[] menuEntries) {
-    ArrayList<MenuEntry> filtered = new ArrayList<>();
-    for (MenuEntry entry : menuEntries) {
-      String target = Text.standardize(Text.removeTags(entry.getTarget()));
-      String option = Text.standardize(Text.removeTags(entry.getOption()));
-     EntryFromConfig entryFromConfig = new EntryFromConfig(option, target);
-      if (indexOfEntry(this.removeOptions, entryFromConfig, menuEntries) == -1)
-        filtered.add(entry);
-    }
-    return filtered.toArray(new MenuEntry[0]);
-  }
-
-
-  @Subscribe
-  public void onFocusChanged(FocusChanged event) {
-    if (!event.isFocused()) {
-      holdingShift = false;
-    }
-
-  }
-
-  private boolean isBankInterfaceClosed() {
-    Widget widgetBankTitleBar = client.getWidget(WidgetInfo.BANK_TITLE_BAR);
-    Widget widgetDepositBox = client.getWidget(12582912);
-    Widget coxPublicChest = client.getWidget(550, 1);
-    Widget coxPrivateChest = client.getWidget(271, 1);
-    return (widgetBankTitleBar == null || widgetBankTitleBar.isHidden()) &&
-        (widgetDepositBox == null || widgetDepositBox.isHidden()) && (coxPublicChest == null
-        || coxPublicChest.isHidden()) && (coxPrivateChest == null || coxPrivateChest.isHidden());
-  }
-
-  static {
-    EXAMINE_TYPES = ImmutableSet
-        .of(MenuAction.EXAMINE_ITEM, MenuAction.EXAMINE_ITEM_GROUND, MenuAction.EXAMINE_NPC,
-            MenuAction.EXAMINE_OBJECT);
-    SPLITTER = Splitter.on("\n").omitEmptyStrings().trimResults();
-  }
-
-  private void specialSwap(int actionId, int opId) {
-    MenuEntry[] menuEntries = client.getMenuEntries();
-
-    for (int i = menuEntries.length - 1; i >= 0; --i) {
-      MenuEntry entry = menuEntries[i];
-      if (entry.getType() == actionId && entry.getIdentifier() == opId) {
-        entry.setType(MenuAction.CC_OP.getId());
-        menuEntries[i] = menuEntries[menuEntries.length - 1];
-        menuEntries[menuEntries.length - 1] = entry;
-        client.setMenuEntries(menuEntries);
-        break;
-      }
-    }
-  }
-
-  public void keyTyped(KeyEvent event) {
-  }
-
-  public void keyPressed(KeyEvent event) {
-    if (event.getKeyCode() == 16) {
-      holdingShift = true;
-    }
-
-  }
-
-  public void keyReleased(KeyEvent event) {
-    if (event.getKeyCode() == 16) {
-      holdingShift = false;
-    }
-
   }
 }
